requirements.txt
aiohttp==3.6.3
alembic==1.4.3
async-timeout==3.0.1
attrs==20.3.0
banal==1.0.1
boto3==1.16.47
botocore==1.19.47
cffi==1.14.4
chardet==3.0.4
dataset==1.4.3
discord.py==1.5.1
idna==3.0
jmespath==0.10.0
Mako==1.1.3
MarkupSafe==1.1.1
multidict==4.7.6
pycparser==2.20
PyNaCl==1.3.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.22
urllib3==1.26.2
yarl==1.5.1
